test: add endpoint tests for names functionality using PostgreSQL
