+++ nicknamer/server/src/name.rs
@@ -172,114 +172,3 @@ async fn names_handler(
 pub fn create_name_router() -> Router {
     Router::new().route("/names", get(names_handler))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::body::Body;
-    use axum::http::Request;
-    use migration::MigratorTrait;
-    use sea_orm::Database;
-    use tower::ServiceExt;
-
-    /// Test helper to create a test database connection.
-    async fn create_test_db() -> DatabaseConnection {
-        let db = Database::connect("sqlite::memory:").await.unwrap();
-        migration::Migrator::up(&db, None).await.unwrap();
-        db
-    }
-
-    /// Test helper to create test names in the database.
-    async fn create_test_names(db: &DatabaseConnection) -> Vec<Name> {
-        let name_service = NameService::new(db);
-
-        let name1 = name_service
-            .create_name(123456789, "TestUser1".to_string())
-            .await
-            .unwrap();
-
-        let name2 = name_service
-            .create_name(987654321, "TestUser2".to_string())
-            .await
-            .unwrap();
-
-        vec![name1, name2]
-    }
-
-    #[tokio::test]
-    async fn can_display_names_table_when_names_exist() {
-        let db = create_test_db().await;
-        let _test_names = create_test_names(&db).await;
-
-        let app = create_name_router().layer(Extension(Arc::new(db)));
-
-        let request = Request::builder()
-            .uri("/names")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .unwrap();
-        let body_text = std::str::from_utf8(&body).unwrap();
-
-        // Check that the page contains the expected content
-        assert!(body_text.contains("Names Database"));
-        assert!(body_text.contains("Stored Names"));
-        assert!(body_text.contains("TestUser1"));
-        assert!(body_text.contains("TestUser2"));
-        assert!(body_text.contains("123456789"));
-        assert!(body_text.contains("987654321"));
-        assert!(body_text.contains("Total Names"));
-        assert!(body_text.contains("2")); // Should show count of 2 names
-    }
-
-    #[tokio::test]
-    async fn can_display_empty_names_table_when_no_names_exist() {
-        let db = create_test_db().await;
-
-        let app = create_name_router().layer(Extension(Arc::new(db)));
-
-        let request = Request::builder()
-            .uri("/names")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .unwrap();
-        let body_text = std::str::from_utf8(&body).unwrap();
-
-        // Check that the page shows empty state
-        assert!(body_text.contains("Names Database"));
-        assert!(body_text.contains("No names found in the database"));
-    }
-
-    #[tokio::test]
-    async fn names_endpoint_returns_correct_content_type() {
-        let db = create_test_db().await;
-
-        let app = create_name_router().layer(Extension(Arc::new(db)));
-
-        let request = Request::builder()
-            .uri("/names")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(
-            response.headers().get("content-type").unwrap(),
-            "text/html; charset=utf-8"
-        );
-    }
-}
