+++ nicknamer/server/tests/names_endpoints_tests.rs
@@ -1,0 +1,120 @@
+use axum::body::Body;
+use axum::extract::Extension;
+use axum::http::{Request, StatusCode};
+use nicknamer_server::name::{Name, NameService, create_name_router};
+use sea_orm::DatabaseConnection;
+use std::sync::Arc;
+use testcontainers_modules::{postgres, testcontainers};
+use tower::ServiceExt;
+
+mod common;
+
+/// Test context for endpoint tests.
+pub struct TestContext {
+    #[allow(dead_code)] // container is kept to ensure it's not dropped
+    pub container: testcontainers::ContainerAsync<postgres::Postgres>,
+    pub db: DatabaseConnection,
+}
+
+/// Setup function for endpoint tests using PostgreSQL container.
+async fn setup() -> anyhow::Result<TestContext> {
+    // Allow multiple calls to init for tests.
+    let _ = tracing_subscriber::fmt().try_init();
+    let container = common::setup_container().await?;
+    let db = common::setup_db(&container).await?;
+    Ok(TestContext { db, container })
+}
+
+/// Test helper to create test names in the database.
+async fn create_test_names(db: &DatabaseConnection) -> Vec<Name> {
+    let name_service = NameService::new(db);
+
+    let name1 = name_service
+        .create_name(123456789, "TestUser1".to_string())
+        .await
+        .unwrap();
+
+    let name2 = name_service
+        .create_name(987654321, "TestUser2".to_string())
+        .await
+        .unwrap();
+
+    vec![name1, name2]
+}
+
+#[tokio::test]
+async fn can_display_names_table_when_names_exist() {
+    let state = setup().await.expect("Failed to setup test context");
+    let _test_names = create_test_names(&state.db).await;
+
+    let app = create_name_router().layer(Extension(Arc::new(state.db)));
+
+    let request = Request::builder()
+        .uri("/names")
+        .body(Body::empty())
+        .unwrap();
+
+    let response = app.oneshot(request).await.unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = axum::body::to_bytes(response.into_body(), usize::MAX)
+        .await
+        .unwrap();
+    let body_text = std::str::from_utf8(&body).unwrap();
+
+    // Check that the page contains the expected content
+    assert!(body_text.contains("Names Database"));
+    assert!(body_text.contains("Stored Names"));
+    assert!(body_text.contains("TestUser1"));
+    assert!(body_text.contains("TestUser2"));
+    assert!(body_text.contains("123456789"));
+    assert!(body_text.contains("987654321"));
+    assert!(body_text.contains("Total Names"));
+    assert!(body_text.contains("2")); // Should show count of 2 names
+}
+
+#[tokio::test]
+async fn can_display_empty_names_table_when_no_names_exist() {
+    let state = setup().await.expect("Failed to setup test context");
+
+    let app = create_name_router().layer(Extension(Arc::new(state.db)));
+
+    let request = Request::builder()
+        .uri("/names")
+        .body(Body::empty())
+        .unwrap();
+
+    let response = app.oneshot(request).await.unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = axum::body::to_bytes(response.into_body(), usize::MAX)
+        .await
+        .unwrap();
+    let body_text = std::str::from_utf8(&body).unwrap();
+
+    // Check that the page shows empty state
+    assert!(body_text.contains("Names Database"));
+    assert!(body_text.contains("No names found in the database"));
+}
+
+#[tokio::test]
+async fn names_endpoint_returns_correct_content_type() {
+    let state = setup().await.expect("Failed to setup test context");
+
+    let app = create_name_router().layer(Extension(Arc::new(state.db)));
+
+    let request = Request::builder()
+        .uri("/names")
+        .body(Body::empty())
+        .unwrap();
+
+    let response = app.oneshot(request).await.unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+    assert_eq!(
+        response.headers().get("content-type").unwrap(),
+        "text/html; charset=utf-8"
+    );
+}
